get handler for server

+++ src/main.rs
@@ -9,7 +9,7 @@ pub mod utils;
 use bot::Bot;
 use tokio::signal;
 
-pub const PORT: u32 = 65005;
+pub const PORT: usize = 65005;
 
 #[tokio::main]
 async fn main() {
